+++ src/main/java/com/develhope/spring/vehicles/optionals/enums/TypeOfRim.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.vehicles.optionals.enums;
+
+public enum TypeOfRim {
+     CHROME,
+     COLOR,
+    AlloyWheels;
+}
